squashing ommits

+++ secureli/modules/language_analyzer/language_support.py
@@ -37,12 +37,14 @@ class LanguageSupportService:
         languages: list[str],
         language_config_result: language.BuildConfigResult,
         overwrite_pre_commit: bool,
+        preserve_precommit_config: bool = False,
     ) -> language.LanguageMetadata:
         """
         Applies Secure Build support for the provided languages
         :param languages: list of languages to provide support for
         :param language_config_result: resulting config from language hook detection
         :param overwrite_pre_commit: flag to determine if config should overwrite or append to config file
+        :param preserve_precommit_config: If true, preserve the existing pre-commit configuration
         :raises LanguageNotSupportedError if support for the language is not provided
         :return: Metadata including version of the language configuration that was just installed
         as well as a secret-detection hook ID, if present.
@@ -58,14 +60,15 @@ class LanguageSupportService:
             language_config_result.linter_configs
         )
 
-        pre_commit_file_mode = "w" if overwrite_pre_commit else "a"
-        with open(path_to_pre_commit_file, pre_commit_file_mode) as f:
-            data = (
-                language_config_result.config_data
-                if overwrite_pre_commit
-                else language_config_result.config_data["repos"]
-            )
-            f.write(yaml.dump(data))
+        if not preserve_precommit_config:
+            pre_commit_file_mode = "w" if overwrite_pre_commit else "a"
+            with open(path_to_pre_commit_file, pre_commit_file_mode) as f:
+                data = (
+                    language_config_result.config_data
+                    if overwrite_pre_commit
+                    else language_config_result.config_data["repos"]
+                )
+                f.write(yaml.dump(data))
 
         # Add .secureli/ to the gitignore folder if needed
         self.git_ignore.ignore_secureli_files()
